Initial build separation. Waiting to refactor tests
