+++ gtclang/test/unit-test/Frontend/CMakeLists.txt
@@ -19,13 +19,12 @@ set(test_name ${PROJECT_NAME}UnittestFrontend)
 add_executable(${test_name}
   TestParsing.cpp
   TestPreprocessing.cpp
-  TestIRSplitting.cpp
   TestSIRFrontend.cpp
   TestMain.cpp
 )
 
 target_add_gtclang_standard_props(${test_name})
-target_link_libraries(${test_name} ${PROJECT_NAME} ${PROJECT_NAME}Unittest DawnUnittest gtest)
+target_link_libraries(${test_name} ${PROJECT_NAME} ${PROJECT_NAME}Unittest gtest)
 
 set_target_properties(${test_name} PROPERTIES
   RUNTIME_OUTPUT_DIRECTORY ${CMAKE_BINARY_DIR}/bin/unittest
